factory: bug fix + split checkAdaptation in new Manifest
