+++ src/manifest/period.ts
@@ -13,14 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 import arrayFind from "array-find";
+import { Subject } from "rxjs";
+import { ICustomError } from "../errors";
+import MediaError from "../errors/MediaError";
+import log from "../log";
 import { CustomRepresentationFilter } from "../net/types";
+import arrayIncludes from "../utils/array-includes";
 import generateNewId from "../utils/id";
 import { normalize as normalizeLang } from "../utils/languages";
 import Adaptation, {
   IAdaptationArguments,
   IAdaptationType,
+  SUPPORTED_ADAPTATIONS_TYPE,
 } from "./adaptation";
 import { StaticRepresentationIndex } from "./representation_index";
 
@@ -63,6 +68,7 @@ export default class Period {
    */
   constructor(
     args : IPeriodArguments,
+    warning$: Subject<Error|ICustomError>,
     customRepresentationFilter? : CustomRepresentationFilter
   ) {
     this.id = args.id;
@@ -72,13 +78,34 @@ export default class Period {
           if (args.adaptations[type]) {
             const adaptationsForType = args.adaptations[type];
             if (adaptationsForType) {
-              acc[type] = adaptationsForType.map((adaptation) => {
-                return new Adaptation(adaptation, customRepresentationFilter);
-              });
+              acc[type] = adaptationsForType
+                .filter((adaptation) => {
+                  if (!arrayIncludes(SUPPORTED_ADAPTATIONS_TYPE, adaptation.type)) {
+                    log.info("not supported adaptation type", adaptation.type);
+                    warning$.next(
+                      new MediaError("MANIFEST_UNSUPPORTED_ADAPTATION_TYPE", null, false)
+                    );
+                    return false;
+                  } else {
+                    return true;
+                  }
+                })
+                .map((adaptation) => {
+                  return new Adaptation(adaptation, warning$, customRepresentationFilter);
+                })
+                .filter((adaptation) => adaptation.representations.length);
             }
           }
           return acc;
         }, {});
+
+    if (
+      (!this.adaptations.video || !this.adaptations.video.length) &&
+      (!this.adaptations.audio || !this.adaptations.audio.length)
+    ) {
+      throw new MediaError("MANIFEST_PARSE_ERROR", null, true);
+    }
+
     this.duration = args.duration;
     this.start = args.start;
